Server-setup/python_scripts/requirements.txt
requests
gitpython
pygithub